frontend/api-deployment/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
qdrant-client==1.7.0
datasets==2.16.1
numpy==1.26.3
python-dotenv==1.0.0
